+++ Fay/Source/Fay/Entity/Components.h
@@ -24,16 +24,17 @@ namespace Fay
 
 		TransformComponent() = default;
 		TransformComponent(const Vec3& pos, const Vec3& rot, const Vec3& translation, const Vec3& scale) : position(pos),
-			rotation(rot), translation(translation), scale(scale){}
+			rotation(rot), translation(translation), scale(scale) {
+		}
 	};
 
-	struct CollisionSpriteComponent
+	struct CollisionComponent
 	{
 		Vec3 pos;
 		Vec3 size;
-
-		CollisionSpriteComponent() = default;
-		CollisionSpriteComponent(const Vec3& pos, const Vec3& size) : pos(pos), size(size) {}
+		
+		CollisionComponent() = default;
+		CollisionComponent(const Vec3& pos, const Vec3& size) : pos(pos), size(size) {}
 	};
 	template<typename... Components>
 	struct ComponentGroup
@@ -94,21 +95,39 @@ namespace Fay
 	struct CubeComponent
 	{
 		Renderable* cube = nullptr; 
-		Vec3 position = Vec3(0, 0, 0);
-		Vec3 size = Vec3(0, 0, 0);
-		Vec4 color = Vec4(0, 0, 0, 0);
-		bool hasCol = false;
 		//Mat4 modelMatrix;
 		CubeComponent() = default;
-		CubeComponent(Renderable* c) : cube(c) 
+		CubeComponent(Renderable* c) : cube(c) {}
+		void setCollision(bool cond)
 		{
-			//modelMatrix = Mat4::translation(c->getPosition()) * Mat4::scale(c->getSize());
-			if (cube)
-			{
-				position = cube->getPosition();
-				size = cube->getSize();
-				color = cube->getColor();
-			}
+			//sprite->setColision(cond);
+			cube->setCollision(cond);
+		}
+		void setColor(Vec4 color)
+		{
+			cube->setColor(color);
+		}
+		void setPosition(Vec3 position)
+		{
+			cube->setPosition(position);
+		}
+		void setSize(Vec3 size)
+		{
+			cube->setSize(size);
+		}
+		bool getCollision() const
+		{
+			return cube->getCollision();
+		}
+		Vec3 getPosition() const
+		{
+			return cube->getPosition();
+		}
+		Vec3 getSize() const {
+			return cube->getSize();
+		}
+		Vec4 getColor() const {
+			return cube->getColor();
 		}
 	};
 	/*
@@ -153,7 +172,7 @@ namespace Fay
 		SpriteComponent, 
 		CubeComponent, 
 		ScriptComponent,
-		CollisionSpriteComponent
+		CollisionComponent
 	>;
 
 }