+++ Fay/Source/Fay/Entity/Tile.h
@@ -9,6 +9,8 @@
 namespace Fay
 {
 	class Texture;
+	class Sprite;
+	using EntityID = uint32_t;
 	struct Tile
 	{
 		int id;
@@ -24,12 +26,24 @@ namespace Fay
 		std::string texturePath;
 		Texture* texture = nullptr;
 		Vec4 color;
+		Vec3 size;
+		Vec3 position;
 
-		TileInfo() : id(0), name(""), isTexture(false), texturePath(""), texture(nullptr), color(0.0f, 0.0f, 0.0f, 0.0f) {}
+		TileInfo() : id(0), name(""), isTexture(false), texturePath(""), texture(nullptr), color(0.0f, 0.0f, 0.0f, 0.0f), size(1, 1, 1), position(0, 0, 0){}
 		TileInfo(int id, const std::string& name, const Vec4& color)
 			: id(id), name(name), isTexture(false), texturePath(""), texture(nullptr), color(color) {}
 
 		TileInfo(int id, const std::string& name, const std::string& texPath)
-			: id(id), name(name), isTexture(true), texturePath(texPath), texture(nullptr), color(0.0f, 0.0f, 0.0f, 0.0f) {}
+			: id(id), name(name), isTexture(true), texturePath(texPath), texture(nullptr), color(0.0f, 0.0f, 0.0f, 0.0f) {
+		}
+		void setSize(const Vec3& size)
+		{
+			this->size = size;
+		}
+		const Vec3& getSize() const
+		{
+			return this->size;
+		}
+
 	};
 }