+++ Fay/Source/Fay/Renderer/3DRenderer/Cube.cpp
@@ -2,9 +2,10 @@
 namespace Fay
 {
 		//RenderDimension dimension = RenderDimension::D2
-		Cube::Cube(float x, float y, float z, float width, float height, float depth, const Vec4& color)
-			: Renderable(Vec3(x, y, z), Vec3(width, height, depth), color, RenderDimension::D3), pos(m_position), size(m_size), color(m_color)
+		Cube::Cube(uint32_t& id, float x, float y, float z, float width, float height, float depth, const Vec4& color)
+			: Renderable(id, Vec3(x, y, z), Vec3(width, height, depth), color, RenderDimension::D3), pos(m_position), size(m_size), color(m_color), id(m_id)
 		{
+			m_id = id;
 			m_position = Vec3(x, y, z);
 			m_size = Vec3(width, height, depth);
 		}