+++ Fay/Source/Fay/Renderer/3DRenderer/Cube.h
@@ -5,12 +5,13 @@ namespace Fay
 		class Cube : public Renderable
 		{
 		public: 
+			uint32_t& id;
 			Vec3& pos;
 			Vec3& size;
 			Vec4& color;
 			//bool hasCollision;
 		public:
-			Cube(float x, float y, float z, float width, float height, float depth, const Vec4& color);
+			Cube(uint32_t& id, float x, float y, float z, float width, float height, float depth, const Vec4& color);
 			
 			void setPosition(Vec3& pos);
 			Mat4 getModelMatrix() const {