+++ Fay/Source/Fay/Renderer/Scene.h
@@ -10,15 +10,16 @@
 #include <Entity/Entity.h>
 #include <Entity/ComponentManager.h>
 #include <Entity/Components.h>
+#include <Graphics/Layers/TileLayer.h>
 namespace Fay
 {
 	class Scene
 	{
 	public:
-		void addObject(Renderable* object); // Entity is a single standalone entity that takes either a Sprite* or Cube* and has an id and can be multiple
+		void addObject(Renderable* object);
 		void removeObject(Renderable* object);
 		void clear();
-
+		void render(TileLayer* renderingLayer) const;
 		const std::vector<Renderable*>& getObjects() const { return m_objects; }
 
 		bool saveScene(const std::string& filepath) const;
@@ -30,8 +31,7 @@ namespace Fay
 		void setSceneType(SceneType type);
 		SceneType getSceneType() { return m_ActiveScene; }
 		bool canSwitchScene() const;
-		EntityID generateEntityID() { return m_nextEntityID++; }
-		void resetEntityID() { m_nextEntityID = 0; }
+		EntityID getNextId();
 	private:
 		bool has2DEntities() const;
 		bool has3DEntities() const;