+++ Fay/Source/Fay/Scripting/FayRuntime/API.cs
@@ -7,39 +7,6 @@ using System.Threading.Tasks;
 
 namespace FayRuntime
 {
-    public struct Vector2
-    {
-        public float X, Y;
-
-        public Vector2(float x, float y)
-        {
-            X = x;
-            Y = y;
-        }
-    }
-    public struct Vector3
-    {
-        public float X, Y, Z;
-
-        public Vector3(float x, float y, float z)
-        {
-            X = x;
-            Y = y;
-            Z = z;
-        }
-    }
-    public struct Vector4
-    {
-        public float X, Y, Z, W;
-
-        public Vector4(float x, float y, float z, float w)
-        {
-            X = x;
-            Y = y;
-            Z = z;
-            W = w;
-        }
-    }
     public enum SceneType
     {
         None = 0,
@@ -52,63 +19,70 @@ namespace FayRuntime
 
         public void SetPosition(Vector3 position)
         {
-            InternalCalls.InternalCalls_SetEntityPosition(this, position.X, position.Y, position.Z);
+            InternalCalls.InternalCalls_Entity_SetPosition(this, position.X, position.Y, position.Z);
         }
 
         public Vector3 GetPosition()
         {
-            return InternalCalls.InternalCalls_GetEntityPosition((uint)_entityID);
+            return InternalCalls.InternalCalls_Entity_GetPosition((uint)_entityID);
         }
-        public void SetEntityId(uint id)
+        public void SetId(uint id)
         {
-            InternalCalls.InternalCalls_SetEntityId(this, id);
+            InternalCalls.InternalCalls_Entity_SetID(this, id);
         }
-        public int GetEnityId()
+        public int GetId()
         {
-            return InternalCalls.InternalCalls_GetEntityId(this);
+            return InternalCalls.InternalCalls_Entity_GetID(this);
         }
         public static Entity GetSelected()
         {
             var entity = new Entity();
-            entity.SetEntityId(InternalCalls.InternalCalls_GetSelectedEntity());
+            entity.SetId(InternalCalls.InternalCalls_Entity_GetSelected());
             return entity;
         }
         public bool HasComponent(Type componentType)
         {
-            //bool hasTransform = InternalCalls.InternalCalls_Entity_HasComponent(entity, typeof(FayRuntime.TransformComponent));
             return InternalCalls.InternalCalls_Entity_HasComponent(this, componentType);
         }
-        public void Move(float speed)
+        public void Move(float speed, bool useZ)
         {
-            //if (Input.GetKey(KeyCode.W) && entity.HasComponent(typeof(FayRuntime.TransformComponent)))
-            //{
-            //    Vector3 currentPos = entity.GetPosition();      // Get current position
-            //    float newY = currentPos.Y + 1.0f;               // Increase Y by 1
-            //    entity.SetPosition(new Vector3(currentPos.X, newY, currentPos.Z));  // Set new position
-            //}
-            if(Input.GetKey(KeyCode.W) && this.HasComponent(typeof(FayRuntime.TransformComponent)))
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.W))
             {
                 Vector3 currentPos = this.GetPosition();
                 float newY = currentPos.Y + speed;
                 this.SetPosition(new Vector3(currentPos.X, newY, currentPos.Z));
             }
-            if (Input.GetKey(KeyCode.S) && this.HasComponent(typeof(FayRuntime.TransformComponent)))
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.S))
             {
                 Vector3 currentPos = this.GetPosition();
                 float newY = currentPos.Y - speed;
                 this.SetPosition(new Vector3(currentPos.X, newY, currentPos.Z));
             }
-            if (Input.GetKey(KeyCode.A) && this.HasComponent(typeof(FayRuntime.TransformComponent)))
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.A))
             {
                 Vector3 currentPos = this.GetPosition();
                 float newX = currentPos.X - speed;
                 this.SetPosition(new Vector3(newX, currentPos.Y, currentPos.Z));
             }
-            if (Input.GetKey(KeyCode.D) && this.HasComponent(typeof(FayRuntime.TransformComponent)))
+
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.D))
             {
                 Vector3 currentPos = this.GetPosition();
                 float newX = currentPos.X + speed;
                 this.SetPosition(new Vector3(newX, currentPos.Y, currentPos.Z));
+            }
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.Up) && useZ)
+            {
+                Vector3 currentPos = this.GetPosition();
+                float newZ = currentPos.Z + speed;
+                this.SetPosition(new Vector3(currentPos.X, currentPos.Y, newZ));
+            }
+
+            if (this.HasComponent(typeof(FayRuntime.TransformComponent)) && Input.GetKey(KeyCode.Down) && useZ)
+            {
+                Vector3 currentPos = this.GetPosition();
+                float newZ = currentPos.Z - speed;
+                this.SetPosition(new Vector3(currentPos.X, currentPos.Y, newZ));
             }
         }
         public Entity() { }
@@ -117,6 +91,7 @@ namespace FayRuntime
             _entityID = id;
         }
     }
+    // Components
     public class TransformComponent
     {
         public Vector3 position;
@@ -124,52 +99,13 @@ namespace FayRuntime
         public Vector3 rotation;
         public Vector3 scale;
     }
-    public class Sprite
-    {
-        private IntPtr nativePtr; // Holds the native Sprite* pointer
-        private bool hasColision;
-        public Sprite(IntPtr ptr)
-        {
-            nativePtr = ptr;
-        }
-
-        public Vector3 GetPosition()
-        {
-            return InternalCalls.InternalCalls_Sprite_GetPosition(nativePtr);
-        }
-
-        public void SetPosition(Vector3 position)
-        {
-            InternalCalls.InternalCalls_Sprite_SetPosition(nativePtr, position);
-        }
-
-        //public void SetCollision(IntPtr entity, bool condition)
-        //{
-
-        //}
-    }
-
-    public class Cube
-    {
-        private IntPtr nativePtr; // Native pointer that points to the actual Cube* pointer must be void* in c++
-
-        public Cube(IntPtr ptr)
-        {
-            nativePtr = ptr;
-        }
-
-        public Vector3 GetPosition()
-        {
-            return InternalCalls.InternalCalls_Cube_GetPosition(nativePtr);
-        }
-
-        public void SetPosition(Vector3 position)
-        {
-            InternalCalls.InternalCalls_Cube_SetPosition(nativePtr, position);
-        }
-    }
-
+   
     public class SpriteComponent
+    {
+        public Vector3 position;
+        public Vector3 scale;
+    }
+    public class CubeComponent
     {
         public Vector3 position;
         public Vector3 scale;
@@ -180,9 +116,9 @@ namespace FayRuntime
         public Vector3 rotation;
         public Vector3 scale;
     }
-    public class CollisionSpriteComponent
+    public class CollisionComponent
     {
         public Vector3 position;
-        public Vector2 size;
+        public Vector3 size;
     }
 }