+++ Fay/Source/Fay/Scripting/FayRuntime/InternalCalls.cs
@@ -9,46 +9,28 @@ namespace FayRuntime
 {
     public static class InternalCalls
     {
-        // Postion stuff
+        // Component
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_SetEntityPosition(object entity, float x, float y, float z);
-
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern Vector3 InternalCalls_GetEntityPosition(uint entityID);
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_SetEntityCollision(int entityID, bool condition);
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern bool InternalCalls_GetEntityCollision(int entityID);
-        // Misc transform component 
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern TransformComponent InternalCalls_GetTransformComponent(Entity entity);
-
-        // Entity stuff
+        internal static extern TransformComponent InternalCalls_Entity_GetTransformComponent(Entity entity);
         [MethodImpl(MethodImplOptions.InternalCall)]
         internal static extern bool InternalCalls_Entity_HasComponent(object entity, Type componentType);
+        // Entity 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern int InternalCalls_GetEntityId(object entity);
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern uint InternalCalls_GetSelectedEntity();
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_SetEntityId(object entity, uint id);
+        internal static extern void InternalCalls_Entity_SetPosition(object entity, float x, float y, float z);
 
-        // Sprite 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern Vector3 InternalCalls_Sprite_GetPosition(IntPtr spritePtr);
+        internal static extern Vector3 InternalCalls_Entity_GetPosition(uint entityID);
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_Sprite_SetPosition(IntPtr spitePtr, Vector3 pos);
-        // Implement in c++
+        internal static extern void InternalCalls_Entity_SetCollision(int entityID, bool condition);
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_Sprite_SetCollision(IntPtr entity, bool condition);
+        internal static extern bool InternalCalls_Entity_GetCollision(int entityID);
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern bool InternalCalls_Sprite_GetCollision(IntPtr entity);
-        // Cube
-        [MethodImpl (MethodImplOptions.InternalCall)]
-        internal static extern Vector3 InternalCalls_Cube_GetPosition(IntPtr cubePtr);
+        internal static extern int InternalCalls_Entity_GetID(object entity);
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern void InternalCalls_Cube_SetPosition(IntPtr cubePtr, Vector3 pos);
-
+        internal static extern int InternalCalls_Entity_SetID(object entity, uint id);
+        [MethodImpl(MethodImplOptions.InternalCall)]
+        internal static extern uint InternalCalls_Entity_GetSelected();
+        // Input handling
         [MethodImpl(MethodImplOptions.InternalCall)]
         internal static extern bool InternalCalls_Window_KeyPressed(int keyCode);
         [MethodImpl(MethodImplOptions.InternalCall)]